requirements.txt
PyQt6==6.6.1
pyscard==2.0.8
PyMySQL==1.1.1
bcrypt==4.1.2
cryptography==44.0.2
cffi==1.17.1
pycparser==2.22 